App strucutre update
